crosschain_backend/pyproject.toml
[tool.poetry]
name = "rubic_crosschain"
version = "0.1.0"
description = ""
authors = ["slycaptcha <slycaptcha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-cors-headers = "^3.7.0"
django-extensions = "^3.1.3"
web3 = "^5.21.0"
Werkzeug = "^2.0.1"
requests-http-signature = "^0.2.0"
djangorestframework = "^3.12.4"
django-debug-toolbar = "^3.2.1"
uvicorn = "^0.14.0"
gunicorn = "^20.1.0"
autopep8 = "^1.5.7"
psycopg2 = "^2.9.1"
gevent = "^21.8.0"
celery = {extras = ["redis"], version = "^5.1.2"}
django-filter = "^2.4.0"
drf-spectacular = "^0.18.0"
django-celery-beat = "^2.2.1"
django-admin-rangefilter = "^0.8.1"
solana = "0.18.0"
borsh-construct = "^0.1.0"
pyTelegramBotAPI = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
